refactor: move InvalidParamError and MissingParamError to validation layer
