+++ src/presentation/errors/index.ts
@@ -1,6 +1,4 @@
 export * from './access-denied-error'
 export * from './email-in-use-error'
-export * from './invalid-param-error'
-export * from './missing-param-error'
 export * from './server-error'
 export * from './unauthorized-error'
