+++ src/validation/validators/password-validation.ts
@@ -1,5 +1,5 @@
+import { InvalidParamError } from '@/validation/errors'
 import { type Validation } from '@/presentation/protocols'
-import { InvalidParamError } from '@/presentation/errors'
 
 export class PasswordValidation implements Validation {
   constructor(private readonly field: string) {}
