+++ src/validation/validators/username-validation.ts
@@ -1,6 +1,6 @@
 import { type UsernameValidator } from '@/validation/protocols'
+import { InvalidParamError } from '@/validation/errors'
 import { type Validation } from '@/presentation/protocols'
-import { InvalidParamError } from '@/presentation/errors'
 
 export class UsernameValidation implements Validation {
   constructor(
